signed/unsigned comparison
git-svn-id: svn://svn.icculus.org/twilight/trunk@3348 d7cf8633-e32d-0410-b094-e92efae38249

+++ hmap/common/bspfile.c
@@ -248,7 +248,7 @@ void	LoadBSPFile (char *filename)
 	LoadFile (filename, (void **)&header);
 
 	// swap the header
-	for (i=0 ; i< sizeof(dheader_t)/4 ; i++)
+	for (i=0 ; i< (int)sizeof(dheader_t)/4 ; i++)
 		((int *)header)[i] = LittleLong ( ((int *)header)[i]);
 
 	if (header->version != BSPVERSION)
